requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
anthropic
python-dotenv
slack-bolt
